+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 **?.?.? Changes:**
 
-* Added 2 new effects:
+* Added 3 new effects:
   * No sprinting: Disables sprinting for all characters, lasts 30 seconds
   * Everyone is Invisible: Every character on the stage becomes invisible, lasts 30 seconds
+  * Revive Dead Characters: Revives all recently killed characters
 
 * Duplicate Random Item Stack:
   * Added config option to blacklist specific items from being duplicated